Modemswitch -> Modeswitch and Dialyzer fix
The main part of this PR is to change how the `usb_modeswitch` command
is called to appease Dialyzer. The logic is changed slightly to call
`System.find_executable/1` since not having `usb_modeswitch` installed
is probably the only error that was causing `System.cmd` to raise
previously.
